all dps samples
