+++ azure-iot-device/samples/async-hub-scenarios/provision_symmetric_key.py
@@ -4,10 +4,15 @@
 # license information.
 # --------------------------------------------------------------------------
 
-import os
 import asyncio
 from azure.iot.device.aio import ProvisioningDeviceClient
+import os
+from azure.iot.device.aio import IoTHubDeviceClient
+from azure.iot.device import Message
+import uuid
 
+
+messages_to_send = 10
 provisioning_host = os.getenv("PROVISIONING_HOST")
 id_scope = os.getenv("PROVISIONING_IDSCOPE")
 registration_id = os.getenv("PROVISIONING_REGISTRATION_ID")
@@ -26,6 +31,34 @@ async def main():
 
     print("The complete registration result is")
     print(registration_result.registration_state)
+
+    if registration_result.status == "assigned":
+        print("Will send telemetry from the provisioned device")
+        device_client = IoTHubDeviceClient.create_from_symmetric_key(
+            symmetric_key=symmetric_key,
+            hostname=registration_result.registration_state.assigned_hub,
+            device_id=registration_result.registration_state.device_id,
+        )
+        # Connect the client.
+        await device_client.connect()
+
+        async def send_test_message(i):
+            print("sending message #" + str(i))
+            msg = Message("test wind speed " + str(i))
+            msg.message_id = uuid.uuid4()
+            msg.correlation_id = "correlation-1234"
+            msg.custom_properties["count"] = i
+            msg.custom_properties["tornado-warning"] = "yes"
+            await device_client.send_message(msg)
+            print("done sending message #" + str(i))
+
+        # send `messages_to_send` messages in parallel
+        await asyncio.gather(*[send_test_message(i) for i in range(1, messages_to_send + 1)])
+
+        # finally, disconnect
+        await device_client.disconnect()
+    else:
+        print("Can not send telemetry from the provisioned device")
 
 
 if __name__ == "__main__":
