+++ azure-iot-device/samples/async-hub-scenarios/provision_symmetric_key_group.py
@@ -10,6 +10,11 @@ import base64
 import hmac
 import hashlib
 from azure.iot.device.aio import ProvisioningDeviceClient
+from azure.iot.device.aio import IoTHubDeviceClient
+from azure.iot.device import Message
+import uuid
+
+messages_to_send = 5
 
 provisioning_host = os.getenv("PROVISIONING_HOST")
 id_scope = os.getenv("PROVISIONING_IDSCOPE")
@@ -47,16 +52,24 @@ def derive_device_key(device_id, group_symmetric_key):
 
 # derived_device_key has been computed already using the helper function somewhere else
 # AND NOT on this sample. Do not use the direct master key on this sample to compute device key.
-
-
 derived_device_key_1 = "some_value_already_computed"
 derived_device_key_2 = "some_value_already_computed"
 derived_device_key_3 = "some_value_already_computed"
 
-
 device_ids_to_keys[device_id_1] = derived_device_key_1
 device_ids_to_keys[device_id_2] = derived_device_key_2
 device_ids_to_keys[device_id_3] = derived_device_key_3
+
+
+async def send_test_message(i, client):
+    print("sending message # {index} for client with id {id}".format(index=i, id=client.id))
+    msg = Message("test wind speed " + str(i))
+    msg.message_id = uuid.uuid4()
+    msg.correlation_id = "correlation-1234"
+    msg.custom_properties["count"] = i
+    msg.custom_properties["tornado-warning"] = "yes"
+    await client.send_message(msg)
+    print("done sending message # {index} for client with id {id}".format(index=i, id=client.id))
 
 
 async def main():
@@ -73,10 +86,50 @@ async def main():
     results = await asyncio.gather(
         register_device(device_id_1), register_device(device_id_2), register_device(device_id_3)
     )
-    for index in range(0, len(device_ids_to_keys)):
+
+    clients_to_device_ids = {}
+
+    for index in range(0, len(results)):
         registration_result = results[index]
         print("The complete state of registration result is")
         print(registration_result.registration_state)
+
+        if registration_result.status == "assigned":
+            device_id = registration_result.registration_state.device_id
+
+            print(
+                "Will send telemetry from the provisioned device with id {id}".format(id=device_id)
+            )
+            device_client = IoTHubDeviceClient.create_from_symmetric_key(
+                symmetric_key=device_ids_to_keys[device_id],
+                hostname=registration_result.registration_state.assigned_hub,
+                device_id=registration_result.registration_state.device_id,
+            )
+            # Assign the Id just for print statements
+            device_client.id = device_id
+
+            clients_to_device_ids[device_id] = device_client
+
+        else:
+            print("Can not send telemetry from the provisioned device")
+
+    # connect all the clients
+    await asyncio.gather(*[client.connect() for client in clients_to_device_ids.values()])
+
+    # send `messages_to_send` messages in parallel.
+    await asyncio.gather(
+        *[
+            send_test_message(i, client)
+            for i, client in [
+                (i, client)
+                for i in range(1, messages_to_send + 1)
+                for client in clients_to_device_ids.values()
+            ]
+        ]
+    )
+
+    # disconnect all the clients
+    await asyncio.gather(*[client.disconnect() for client in clients_to_device_ids.values()])
 
 
 if __name__ == "__main__":
