+++ azure-iot-device/samples/sync-samples/provision_symmetric_key_with_payload.py
@@ -5,6 +5,9 @@
 # --------------------------------------------------------------------------
 import os
 from azure.iot.device import ProvisioningDeviceClient
+from azure.iot.device import IoTHubDeviceClient, Message
+import uuid
+import time
 
 
 class Wizard(object):
@@ -39,3 +42,34 @@ print("The status was :-")
 print(registration_result.status)
 print("The etag is :-")
 print(registration_result.registration_state.etag)
+
+
+if registration_result.status == "assigned":
+    print("Will send telemetry from the provisioned device")
+    # Create device client from the above result
+    device_client = IoTHubDeviceClient.create_from_symmetric_key(
+        symmetric_key=symmetric_key,
+        hostname=registration_result.registration_state.assigned_hub,
+        device_id=registration_result.registration_state.device_id,
+    )
+
+    # Connect the client.
+    device_client.connect()
+
+    for i in range(1, 6):
+        print("sending message #" + str(i))
+        device_client.send_message("test payload message " + str(i))
+        time.sleep(1)
+
+    for i in range(6, 11):
+        print("sending message #" + str(i))
+        msg = Message("test wind speed " + str(i))
+        msg.message_id = uuid.uuid4()
+        msg.custom_properties["tornado-warning"] = "yes"
+        device_client.send_message(msg)
+        time.sleep(1)
+
+        # finally, disconnect
+        device_client.disconnect()
+else:
+    print("Can not send telemetry from the provisioned device")
