+++ azure-iot-device/samples/sync-samples/provision_x509.py
@@ -5,6 +5,10 @@
 # --------------------------------------------------------------------------
 import os
 from azure.iot.device import ProvisioningDeviceClient, X509
+import time
+from azure.iot.device import IoTHubDeviceClient, Message
+import uuid
+
 
 provisioning_host = os.getenv("PROVISIONING_HOST")
 id_scope = os.getenv("PROVISIONING_IDSCOPE")
@@ -27,3 +31,33 @@ registration_result = provisioning_device_client.register()
 
 # The result can be directly printed to view the important details.
 print(registration_result)
+
+if registration_result.status == "assigned":
+    print("Will send telemetry from the provisioned device")
+    # Create device client from the above result
+    device_client = IoTHubDeviceClient.create_from_x509_certificate(
+        x509=x509,
+        hostname=registration_result.registration_state.assigned_hub,
+        device_id=registration_result.registration_state.device_id,
+    )
+
+    # Connect the client.
+    device_client.connect()
+
+    for i in range(1, 6):
+        print("sending message #" + str(i))
+        device_client.send_message("test payload message " + str(i))
+        time.sleep(1)
+
+    for i in range(6, 11):
+        print("sending message #" + str(i))
+        msg = Message("test wind speed " + str(i))
+        msg.message_id = uuid.uuid4()
+        msg.custom_properties["tornado-warning"] = "yes"
+        device_client.send_message(msg)
+        time.sleep(1)
+
+    # finally, disconnect
+
+else:
+    print("Can not send telemetry from the provisioned device")
